Move PlatformHelperVersionStatus to semantic_version
